cm2bm_msg: add update auto reset timeout msg
Adds a cm2bm message allowing an auto reset timeout to be set in BMC fw.
This will define a countdown during which BMC fw expects telemetry
heartbeat to be updated by the chip. If the timer expires, then the
STM32 will step in and reset the chip.

A default auto reset timeout is included in the FW table.

Signed-off-by: Petra Alexson <palexson@tenstorrent.com>