tenstorrent: bh_chip: add chip heartbeat check and auto reset timer
Adds a handler to receive chip's telemetry heartbeat as an indicator
of life.

Adds an auto reset timer to the chip that is restarted with
each successful telemetry heartbeat check.

Adds a reset reason to BM static info that is sent to the
chip after it is reset.

Stores current telemetry heartbeat, auto reset timeout,
and reset reason as part of bh_chip_data.

If communcation with the chip is lost and an auto reset occurs, then set the fan
to max speed.

Signed-off-by: Petra Alexson <palexson@tenstorrent.com>