Updated reset logic to handle two back to back perst asserts
Now that the work loop is entirely synchronous there is no longer a case
where the workarond is not applied once the main loop is reached. Also
considering the case where a second reset interrupt is recieved I
updated the logic to only assert the bus clear flag if trigger_reset has
been cleared.

Signed-off-by: Daniel Rosen <drosen@tenstorrent.com>